requirements.txt
# ──────────────────────────────────────────────────────────────────
# AI Meeting Assistant — Python dependencies
#
# Standard install (CPU inference only):
#   pip install -r requirements.txt
#
# For GPU-accelerated LLM inference on CUDA 12.1 (recommended):
#   Step 1: pip install -r requirements.txt
#   Step 2: pip install llama-cpp-python \
#               --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121 \
#               --force-reinstall \
#               --no-cache-dir
#
# faster-whisper uses a separate CUDA path managed by torch/ctranslate2
# and does not require a special index URL.
# ──────────────────────────────────────────────────────────────────

# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# Speech-to-text (faster-whisper)
faster-whisper>=1.0.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LLM inference — installs CPU wheel by default.
# Replace with the CUDA wheel using the command in the header above.
llama-cpp-python>=0.2.0
